% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{getBioclimVars}
\alias{getBioclimVars}
\title{getBioclimVars}
\usage{
getBioclimVars(occurrences, which.biovars=c(1,12))
}
\arguments{
\item{occurrences}{A matrix or data.frame with three columns and rows to represent idividuals. The first column is either the species name or the age of the fossil.}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (www.XX).}
}
\value{
Returns a data.frame with the original occurrences input appended with columns of bioclimate variables specified. If fossils are included, the returned bioclimate variables are from the closest 1 Ma interval of isotopically scaled climate.
}
\description{
This function retrieves the bioclimate variables described in Nix (1986) for the specified variables and the specified time period. Modern time period uses the Hijmans et al. (2005) high resolution climate interpolations. The time period 10 Ma uses the GCM by XX for the Tortonian and the time period 15 Ma uses the GCM for the Lanhgian by XX. For the one million year intervals outside the modern and past GCMs, the climate was interpolated based on the benthic marine foram stable oxygen isotope ratio curve from XX. The scale of these variables is at a 50 km equidistant point grain size corresponding to Polly XX.
}
\details{
The occurrences argument should contain all extant or all fossils.
}
\author{
A. Michelle Lawing
}
\references{
Hijmans; Polly; GCMs
}

