% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{getEnvelopes}
\alias{getEnvelopes}
\title{getEnvelopes}
\usage{
getEnvelopes(treedata_min,treedata_max,node_est)
}
\description{
This function gets the bioclimate envelopes of species and nodes.
}

