% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{getGeoRate}
\alias{getGeoRate}
\title{getGeoRate}
\usage{
getGeoRate()
}
\arguments{
\item{envelope}{the min and max envelope of each lineage for each time slice}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}
}
\description{
This function calculates the change in suitable habitat through time in geographic space.
}

