% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{getTimeSlice}
\alias{getTimeSlice}
\title{getTimeSlice}
\usage{
getTimeSlice(time, phy, trait, plot.est = TRUE)
}
\arguments{
\item{timeSlice}{single numeric or a vector with the time (or times) to extract the estimated ancestor reconstructions.}

\item{tree}{an object of the class "phylo" that should be dated}

\item{trait}{a vector of both tip values and node estimates that correspond to tree}

\item{plot.est}{a conditional stating whether or not to plot the results}
}
\value{
edge for each time specified, a vector of edges that are present during that time are returned

est for each time specified, a vector of estimates of the ancestral reconstruction along each edge
}
\description{
This function extracts estimated ancestral reconstructions for continuous characters any time specified along a phylogeny for all lineages present at the specified time.
}
\details{
The estimated reconstruction relies on an interpolation between node or between tip and node estimates of the trait. This method assumes a constant rate of evolution along the lineage where the interpolation is taking place.
}
\author{
A. Michelle Lawing
}

