% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{node.estimate.fossils}
\alias{node.estimate.fossils}
\title{node.estimate.fossils}
\usage{
node.estimate.fossils()
}
\arguments{
\item{treedata_min}{tree data object with min estimate of the climate envelope}

\item{treedata_max}{tree data object with max estimate of the climate envelope}

\item{fossils}{the estimate of the climate envelope of the fossil occurrences}

\item{fossils.edges}{the edge number that the fossil occurs on}
}
\description{
To estimate nodes with the placement of fossils on randomly assigned or specifed edges on a tree.
}

