% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{node.estimate}
\alias{node.estimate}
\title{node.estimate}
\usage{
node.estimate(treedata.obj, traitnum, model = "BM", plot.est = TRUE)
}
\arguments{
\item{treedata.obj}{an object of the class "treedata".}

\item{traitnum}{the column number of the trait within the treedata object to be reconstructed.}

\item{model}{the model of evolution to use in the ancestral state reconstruction. Options are c("estimate","BM","OU","EB","trend","lambda","kappa","delta","drift","white").}

\item{plot.est}{whether or not to plot the traitgram of the estimated ancester states.}

\item{...}{arguments to pass to fitContinuous}
}
\value{
an object of the class "node.estimate".

model    if model = "estimate", the best fit model of evolution. If the model was specified, then model is the specified model.

est      the ancestral node estimates of the continuous character.

phy      the phylogeny used for the estimates, which might be transformed depending on the evolutionary model.

BM       if model = "BM", returned values from the fitContinuous function where the model is "BM"

OU       if model = "OU", returned values from the fitContinuous function where the model is "OU"

EB       if model = "EB", returned values from the fitContinuous function where the model is "EB"

trend    if model = "trend", returned values from the fitContinuous function where the model is "trend"

lambda   if model = "lambda", returned values from the fitContinuous function where the model is "lambda"

kappa    if model = "kappa", returned values from the fitContinuous function where the model is "kappa"

delta    if model = "delta", returned values from the fitContinuous function where the model is "delta"

drift    if model = "drift", returned values from the fitContinuous function where the model is "drift"

white    if model = "white", returned values from the fitContinuous function where the model is "white"

fitted   if model = "estimate", returned values from the best fit model of evolution.
}
\description{
This function estimates the ancestral character states for continuous characters given a model of evolution or using the best fit model of evolution from the fitContinuous function in the geiger pacakage. The ancestral states are estimated using GLS described in Martins and Hansen (1997).
}
\details{
See the fitContinuous details for descriptions of the models of evolution and their parameters being estimated.
}
\author{
A. Michelle Lawing
if model is trend, tree should NOT be ultrametric
if model is trend, tree should NOT be ultrametric
if model is drift, tree should NOT be ultrametric
}
\references{
Martins, E. P. and Hansen, T. F. (1997) Phylogenies and the comparative method: a general approach to incorporating phylogenetic information into the analysis of interspecific data. American Naturalist, 149, 646–667.
}
\seealso{
fitContinuous
}

