% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\docType{data}
\name{occurrences}
\alias{occurrences}
\title{Sceloporus occurrence data}
\format{
'occurrences'
A data frame of Sceloporus occurrence records
\describe{
 \item{Species}{Species name}
 \item{Longitude}{Longitude of occurrence}
 \item{Latitude}{Latitude of occurence}
}
}
\source{
<https://www.journals.uchicago.edu/doi/10.1086/687202>
}
\usage{
occurrences
}
\description{
Occurrences for Sceloporus, as collected for Lawing et al 2016. Occurrence records from GBIF, and vetted using expert range maps from IUCN. See reference for further details
Lawing et al (2015) Including fossils in phylogenetic climate reconstructions: A deep time perspective on the climatic niche evolution and diversification of spiny lizards (Sceloporus)
}
\keyword{datasets}
