% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnimatedPPGM.R
\name{plotAnimatedPPGM}
\alias{plotAnimatedPPGM}
\title{plotAnimatedPPGM}
\usage{
plotAnimatedPPGM(envelope, tree, filename="ppgm.gif", which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL)
}
\arguments{
\item{envelope}{the min and max envelope of each lineage for each time slice}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}

\item{filename}{desired filename of output}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (https://www.worldclim.org/data/bioclim.html).}

\item{path}{path to the directory where the results should be saved}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}
}
\value{
An animated gif of species through time
}
\description{
This function creates an animated gif showing the change in modelled suitable habitat through time in geographic space.
}
\details{
Requires ImageMagick or GraphicsMagick to be installed on the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard, Maria-Aleja Hurtado-Materon
}
