% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBumpChart.R
\name{plotBumpChart}
\alias{plotBumpChart}
\title{plotBumpChart}
\usage{
plotBumpChart(sp_data_mean, which.biovars, path="")
}
\arguments{
\item{which.biovars}{a vector with the biovars (see www.worldclim.org for a list of biovars).}

\item{path}{path to the directory where the results should be saved}
}
\value{
bump chart of mean species-climate associations
}
\description{
plotBumpChart
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
