% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraitGram.R
\name{plotTraitGram}
\alias{plotTraitGram}
\title{plotTraitGram}
\usage{
plotTraitGram(treedata_min, treedata_max, node_est, fossils=FALSE, 
which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL,
layerAge=c(0:20))
}
\arguments{
\item{treedata_min}{a tree data object with the min estimate of the climate envelope}

\item{treedata_max}{a tree data object with the max estimate of the climate envelope}

\item{node_est}{the estimate of all the nodes, both min and max}

\item{fossils}{a matrix with four columns of min age, max age, longitude, and latitude, in that order, and rows that are entries for fossil occurrences.}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (https://www.worldclim.org/data/bioclim.html).}

\item{path}{path to the directory where the results should be saved}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}

\item{layerAge}{vector with the ages of the paleoclimate dataframes, if using user submitted paleoclimate data}
}
\value{
a trait gram for minimum and maximum of biovariables
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotraitgram in a specified directory.
}
\examples{
data(sampletrees)
sampletrees <- sample(sampletrees,5)
data(occurrences)
occu <- getBioclimVars(occurrences, which.biovars=c(1,2))
sp_data_min<-sapply(4:5,function(x) tapply(occu[,x],occu$Species,min)) 
sp_data_max<-sapply(4:5,function(x) tapply(occu[,x],occu$Species,max))
ex_min <- geiger::treedata(sampletrees[[1]], sp_data_min, sort=TRUE)
ex_max <- geiger::treedata(sampletrees[[1]], sp_data_max, sort=TRUE)
colnames(ex_min$data)<- colnames(ex_max$data)<-c("bio1","bio2")  #labels biovars
\donttest{nodeest<- nodeEstimateEnvelopes(treedata_min=ex_min,treedata_max=ex_max, 
model="BM",which.biovars=c(1,2),
bounds=list(sigsq = c(min = 0, max = 1000000)))
plotTraitGram(ex_min, ex_max, nodeest$est, which.biovars=c(1,2), 
path=paste(tempdir(),"/",sep=""))}
}
\seealso{
plotTraitGramMultiPhylo
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
