% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{plotTraitGramMultiPhylo}
\alias{plotTraitGramMultiPhylo}
\title{plotTraitGramMultiPhylo}
\usage{
plotTraitGramMultiPhylo()
}
\arguments{
\item{treedata_min}{tree data object with min estimate of the climate envelope}

\item{treedata_max}{tree data object with max estimate of the climate envelope}
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotrait gram in a specified directory.
}

