% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampletrees.R
\docType{data}
\name{sampletrees}
\alias{sampletrees}
\title{Sample of Sceloporus phylogenies}
\format{
'sampletrees'
100 trees as class multiPhylo
}
\source{
<https://www.journals.uchicago.edu/doi/10.1086/687202>

<https://karger.com/cgr/article/127/2-4/166/62387/Chromosome-Evolution-and-Diversification-in-North>
}
\usage{
sampletrees
}
\description{
A sample of 100 dated phylogenies from Leache & Sites (2009), trimmed for analysis by Lawing et al (2016).
\itemize{
 \item Lawing et al (2016) Including fossils in phylogenetic climate reconstructions: A deep time perspective on the climatic niche evolution and diversification of spiny lizards (Sceloporus) The American Naturalist 188(2): 133-148
 \item Leache & Sites (2009) Chromosome evolution and diversification in North American Spiny Lizards (Genus Sceloporus) Cytogenetic and Genome Research 127(2-4) 166-181
 }
}
\keyword{datasets}
