% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scel_fossils.R
\docType{data}
\name{scel_fossils}
\alias{scel_fossils}
\title{Sceloporus fossil data}
\format{
'scel_fossils'
A data frame where each row is a single fossil
\describe{
 \item{MinAge}{Minimum age of fossil}
 \item{MaxAge}{Maximum age of fossil}
 \item{Longitude}{Longitude of occurrence}
 \item{Latitude}{Latitude of occurence}
}
}
\source{
<https://www.journals.uchicago.edu/doi/10.1086/687202>
}
\usage{
scel_fossils
}
\description{
Fossil occurrences for Sceloporus, as collected for Lawing et al 2016. Occurrence records from Paleobiology Database. See reference for further details
Lawing et al (2016) Including fossils in phylogenetic climate reconstructions: A deep time perspective on the climatic niche evolution and diversification of spiny lizards (Sceloporus)
}
\keyword{datasets}
