% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineageClimate.R
\name{getLineageClimate}
\alias{getLineageClimate}
\title{getLineageClimate}
\usage{
getLineageClimate(envelope, tree, which.biovars, 
use.paleoclimate=TRUE, paleoclimateUser=NULL)
}
\arguments{
\item{envelope}{the min and max climate envelope of each lineage for each time slice, as outputted by \code{getEnvelopes()}}

\item{tree}{the phylogeny of all species. An object of class phylo}

\item{which.biovars}{a vector of the numbers of the bioclimate variables to be included. The bioclimate variables number correspond to the table at (https://www.worldclim.org/data/bioclim.html).}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}
}
\value{
\code{matchedClim} list of occurrences points for each lineage, for each time slice of paleoclimate data

\code{lineage} list of lineage specific nodes, as output from phangorn::Ancestors
}
\description{
This function calculates the suitable climate for each specific lineage, starting at the tips and going back through time to the root.
}
\details{
Calculates rate of geographic change of all lineages. Outputs both the geographic center change, and the geographic size change.
}
\examples{
data(sampletrees)
data(occurrences)
data(paleoclimate)
occu <- getBioclimVars(occurrences, which.biovars=1)
tree <- sampletrees[[25]]
#species minimum for biovariable 1
sp_data_min<- tapply(occu[,4],occu$Species,min)
#species maximum for biovariable 1
sp_data_max<- tapply(occu[,4],occu$Species,max)
#convert to treedata object
treedata_min <- geiger::treedata(tree,sp_data_min,sort=TRUE,warnings=F) 
treedata_max <- geiger::treedata(tree,sp_data_max,sort=TRUE,warnings=F)
#estimate node values using Brownian Motion
\donttest{full_est <- nodeEstimateEnvelopes(treedata_min,treedata_max)
node_est <- full_est$est #extract only node estimates
#calculate climate envelopes
example_getEnvelopes <- getEnvelopes(treedata_min, treedata_max, node_est)
#calculate lineage specific climate
example_getLinClim <- getLineageClimate(example_getEnvelopes, tree, which.biovars=1)}
}
\seealso{
\code{getEnvelopes()} \code{getGeoRate()}
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard, Maria A. Hurtado-Materon
}
