% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeEstimateEnvelopes.R
\name{nodeEstimateEnvelopes}
\alias{nodeEstimateEnvelopes}
\title{nodeEstimateEnvelopes}
\usage{
nodeEstimateEnvelopes(treedata_min, treedata_max, fossils=FALSE, fossils.edges=FALSE, 
model="BM", bounds=list(), control=list())
}
\arguments{
\item{treedata_min}{tree data object with min estimate of the climate envelope – list where first object is phylogeny, and second object is array of species with climate data variables (species must match)}

\item{treedata_max}{tree data object with max estimate of the climate envelope}

\item{fossils}{the estimate of the climate envelope of the fossil occurrences}

\item{fossils.edges}{the edge number that the fossil occurs on}

\item{model}{the model of evolution to use in the ancestral state reconstruction. Options are "estimate", "BM", "OU", "EB", "lambda", "kappa", "delta".}

\item{bounds}{bounds used for the model, passes to \code{fitContinuous()}, uses default if none specified.}

\item{control}{setting used for optimization of the model likelihood. Passes to \code{fitContinuous()}.}
}
\value{
an object of the class "nodeEstimate".

\code{model}    if model = "estimate", the best fit model of evolution. If the model was specified, then model is the specified model.

\code{est}      the ancestral node estimates of the continuous character.

\code{phy}      the phylogeny used for the estimates, which might be transformed depending on the evolutionary model.

\code{BM}       if model = "BM", returned values from \code{fitContinuous()} where the model is "BM"

\code{OU}       if model = "OU", returned values from \code{fitContinuous()} where the model is "OU"

\code{EB}       if model = "EB", returned values from \code{fitContinuous()} where the model is "EB"

\code{lambda}   if model = "lambda", returned values from \code{fitContinuous()} where the model is "lambda"

\code{kappa}    if model = "kappa", returned values from \code{fitContinuous()} where the model is "kappa"

\code{delta}    if model = "delta", returned values from \code{fitContinuous()} where the model is "delta"

\code{fitted}   if model = "estimate", returned values from the best fit model of evolution.
}
\description{
This function estimates nodes with the placement of fossils on randomly assigned or specified edges on a tree.
}
\details{
function adds fossils to trees according to \code{addFossil()}, then passes to \code{nodeEstimate()}.
}
\examples{
data(sampletrees)
data(occurrences)
data(scel_fossils)
occu <- getBioclimVars(occurrences, which.biovars=1)
sp_data_min<- tapply(occu[,4],occu$Species,min)
sp_data_max<- tapply(occu[,4],occu$Species,max)
ex_min <- geiger::treedata(sampletrees[[1]], sp_data_min)
ex_max <- geiger::treedata(sampletrees[[1]], sp_data_max)
colnames(ex_min$data)<- colnames(ex_max$data)<-"bio1"  #labels biovars
\dontrun{biofossils <- getBioclimVars(scel_fossils,which.biovars=1)
rownames(biofossils)<-paste("fossil",1:length(biofossils[,1]),sep="")
nodeest<- nodeEstimateEnvelopes(treedata_min=ex_min,treedata_max=ex_max, 
model="BM",fossils=biofossils,
bounds=list(sigsq = c(min = 0, max = 1000000)))}
}
\seealso{
\code{nodeEstimate}, \code{fitContinuous}
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
