% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{plotAnimatedPPGM}
\alias{plotAnimatedPPGM}
\title{plotAnimatedPPGM}
\usage{
plotAnimatedPPGM()
}
\arguments{
\item{envelope}{the min and max envelope of each lineage for each time slice}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}
}
\description{
This function creates an animated gif showing the change in modeled suitable habitat through time in geographic space. It require ImageMagick or GraphicsMagick to be previously installed in the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
}

