% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnimatedPPGMMultiPhylo.R
\name{plotAnimatedPPGMMultiPhylo}
\alias{plotAnimatedPPGMMultiPhylo}
\title{plotAnimatedPPGMMultiPhylo}
\usage{
plotAnimatedPPGMMultiPhylo()
}
\arguments{
\item{envelope}{the min and max envelope of each lineage for each time slice}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}

\item{filename}{filename of output}

\item{path}{path to the directory where the results should be saved}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}
}
\description{
This function creates an animated gif showing the change in modeled suitable habitat through time in geographic space. It requires ImageMagick or GraphicsMagick to be previously installed in the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
