% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{plotGeoRates}
\alias{plotGeoRates}
\title{plotGeoRates}
\usage{
plotGeoRates()
}
\value{
Gray background of all pairwise comparisons of change in geo center and area through time. Blue points on top show the sequential change in geo center and expansion/contraction for all lineages.
}

