% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraitGram.R
\name{plotTraitGram}
\alias{plotTraitGram}
\title{plotTraitGram}
\usage{
plotTraitGram(treedata_min, treedata_max, node_est, fossils=FALSE, 
which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL)
}
\arguments{
\item{treedata_min}{a tree data object with the min estimate of the climate envelope}

\item{treedata_max}{a tree data object with the max estimate of the climate envelope}

\item{node_est}{the estimate of all the nodes, both min and max}

\item{fossils}{a matrix with four columns of age to the closest million year integer, longitude, and latitude, in that order, and rows that are entries for fossil occurrences.}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (https://www.worldclim.org/data/bioclim.html).}

\item{path}{path to the directory where the results should be saved}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}
}
\value{
a trait gram for minimum and maximum of biovariables
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotraitgram in a specified directory.
}
\examples{
data(sampletrees)
data(occurrences)
occurrences <- getBioclimVars(occurrences, which.biovars=1)
sp_data_min<- tapply(occurrences[,4],occurrences$Species,min)
sp_data_max<- tapply(occurrences[,4],occurrences$Species,max)
ex_min <- geiger::treedata(sampletrees[[1]], sp_data_min)
ex_max <- geiger::treedata(sampletrees[[1]], sp_data_max)
colnames(ex_min$data)<- colnames(ex_max$data)<-"bio1"  #labels biovars
\dontrun{nodeest<- nodeEstimateFossils(treedata_min=ex_min,treedata_max=ex_max, 
model="BM", bounds=list(sigsq = c(min = 0, max = 1000000)))
plotTraitGram(ex_min, ex_max, nodeest$est, which.biovars=1)}
}
\seealso{
plotTraitGramMultiPhylo
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
