% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_functions.R
\name{plotTraitGram}
\alias{plotTraitGram}
\title{plotTraitGram}
\usage{
plotTraitGram()
}
\arguments{
\item{treedata_min}{a tree data object with the min estimate of the climate envelope}

\item{treedata_max}{a tree data object with the max estimate of the climate envelope}

\item{node_est}{the estimate of all the nodes, both min and max}
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotrait gram in a specified directory.
}

