% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraitGramMultiPhylo.R
\name{plotTraitGramMultiPhylo}
\alias{plotTraitGramMultiPhylo}
\title{plotTraitGramMultiPhylo}
\usage{
plotTraitGramMultiPhylo(treedata_min, treedata_max, node_est, fossils=FALSE, use.paleoclimate=TRUE, paleoclimateUser=NULL, which.biovars, path="", alpha.trans=as.integer(255/(1+log(length(node_est)))))
}
\arguments{
\item{treedata_min}{tree data object with min estimate of the climate envelope}

\item{treedata_max}{tree data object with max estimate of the climate envelope}

\item{node_est}{the estimate of all the nodes, both min and max}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (https://www.worldclim.org/data/bioclim.html).}
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotrait gram in a specified directory.
}
\seealso{
plotTraitGram
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
