% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraitGramMultiPhylo.R
\name{plotTraitGramMultiPhylo}
\alias{plotTraitGramMultiPhylo}
\title{plotTraitGramMultiPhylo}
\usage{
plotTraitGramMultiPhylo(treedata_min, treedata_max, node_est, 
fossils=FALSE, use.paleoclimate=TRUE, paleoclimateUser=NULL, 
which.biovars, path="")
}
\arguments{
\item{treedata_min}{tree data object with min estimate of the climate envelope}

\item{treedata_max}{tree data object with max estimate of the climate envelope}

\item{node_est}{the estimate of all the nodes, both min and max. Must be in format [[trees]][[permut]][2,species,trait]}

\item{fossils}{a matrix with four columns of age to the closest million year integer, longitude, and latitude, in that order, and rows that are entries for fossil occurrences.}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).}

\item{which.biovars}{A vector of the numbers of the bioclimate variables that should be returned. The bioclimate variables number correspond to the Hijmans table at (https://www.worldclim.org/data/bioclim.html).}

\item{path}{path to the directory where the results should be saved}
}
\description{
Combine the node estimates based on random or specified fossil placement and plot them on a phylotrait gram in a specified directory.
}
\details{
plots a traitgram over multiple phylogenetic trees
}
\examples{
data(sampletrees)
data(occurrences)
biooccu <- getBioclimVars(occurrences, which.biovars=1)
sp_data_min<- tapply(biooccu[,4],biooccu$Species,min)
sp_data_max<- tapply(biooccu[,4],biooccu$Species,max)
treedata_min <- treedata_max <- node_est <- envelope <- list()
\dontrun{for (tr in 1:length(sampletrees)){
  treedata_min[[tr]] <- geiger::treedata(sampletrees[[tr]],sp_data_min,sort=TRUE,warnings=F)
  treedata_max[[tr]] <- geiger::treedata(sampletrees[[tr]],sp_data_max,sort=TRUE,warnings=F)
  full_est <- nodeEstimateEnvelopes(treedata_min[[tr]],treedata_max[[tr]])
  node_est[[tr]]<-lapply(1, function(p) full_est$est)
  envelope[[tr]] <- getEnvelopes(treedata_min[[tr]], treedata_max[[tr]], node_est[[tr]])
}
plotTraitGramMultiPhylo(treedata_min,treedata_max, node_est, which.biovars=1)
}
}
\seealso{
plotTraitGram
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard
}
