% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{ppgm}
\alias{ppgm}
\title{ppgm}
\usage{
ppgm(occurrences, fossils, tree, fossils.edges, model = "BM", plot.est=TRUE)
}
\arguments{
\item{occurrences}{a matrix with three columns of species name, longitude, and latitude, in that order, and rows that are entries for species occurrences. The bioclimate variables can be included for each occurrence in following columns. They must be in order 1 through 19.}

\item{fossils}{a matrix with three columns of age to the closest million year integer, longitude, and latitude, in that order, and rows that are entries for fossil occurrences. The bioclimate variables can be included for each occurrence in following columns. They must be in order 1 through 19.}

\item{fossils.edges}{a vector of edges that the fossils belong to. Must be in the same order of the fossils argument. If fossils.edges is false, the the function randomly assigns the location of the fossils depending on the age (see details for more information).}

\item{model}{the model of evolution to use to estimate ancestor nodes. Argument is passed onto to function node.estimate.}

\item{permut}{the number of times to randomly place fossils in phylogeny and estimate ancestor states.}

\item{which.biovars}{a vector with the biovars to include in model (see www.worldclim.org for a list of biovars). If "ALL", then all 19 biovars are included in analysis.}

\item{path}{path to the directory where the results should be saved.}

\item{tree}{phylogeny of species from first column of occurrences argument.}
}
\description{
ppgm makes a paleophylogeographic species distribution model using the bioclimate envelope method for a specified time period. Currently, models are only available for North America.
}
\details{
If the 19 bioclimate variables are not supplied with the occurrences or with the fossils, they will be extracted from the closest 50km point location in the modern or paleoclimate maps that are loaded in with this function. The paleoclimate maps are isotopically scaled between general circulation models (see Lawing and Polly 2011; Rodder et al. 2013) and modern climate (see Hijmans et al. 2005). The fossils paleoclimate data is extracted to the closest million year paleoclimate map. Paleoclimate maps are derived at one million year intervals for the past 20 Ma. The tree (phylogeny) should be dichotomous and the species names should match the names in the first column of the occurrences argument. fix to work with the nice plots on multiple trees, and check and see if it works when no suitable habitat is identified. also, make option to print any animated lineage through time?
}

