% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppgmMESS.R
\name{ppgmMESS}
\alias{ppgmMESS}
\title{ppgmMESS}
\usage{
ppgmMESS(cem_min, cem_max, est, tree, fossils, timeslice, 
which.biovars, path = "", use.paleoclimate=TRUE, paleoclimateUser = NULL, 
which.plot = c("all","mess","none"))
}
\arguments{
\item{cem_min}{the cem min output from the ppgm function. cbind() if there are multiple variables.}

\item{cem_max}{the cem max output from the ppgm function. cbind() if there are multiple variables.}

\item{est}{the node_est output from the ppgm function, in list format. [tree][1][min and max][no.of species]}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}

\item{fossils}{a matrix with four columns of age to the closest million year integer, longitude, and latitude, in that order, and rows that are entries for fossil occurrences.}

\item{timeslice}{the time in million of years ago to project MESS maps (0 to 20). can handle single timeslice or vector of times.}

\item{which.biovars}{the biovariable number(s) between 1 and 19. handles vectors.}

\item{path}{directory where plots should be stored}

\item{use.paleoclimate}{if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided}

\item{paleoclimateUser}{list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19.}

\item{which.plot}{"all" plots trait maps and MESS, "mess" plots MESS map, "none" does not plot}
}
\description{
This creates a MESS map for given time slices, climate envelopes, and paleoclimate models.
}
\details{
plots MESS maps of climate envelope model for specific time slices. Can either plot individual biovariables, or combined.
}
\examples{
data(sampletrees)
data(occurrences)
data(scel_fossils)
biooccu <- getBioclimVars(occurrences, which.biovars=1)
sp_data_min<- tapply(biooccu[,4],biooccu$Species,min)
sp_data_max<- tapply(biooccu[,4],biooccu$Species,max)
treedata_min <- treedata_max <- node_est <- list()
\dontrun{biofossils <- getBioclimVars(scel_fossils,which.biovars=1)
rownames(biofossils)<-paste("fossil",1:length(biofossils[,1]),sep="")
for (tr in 1:length(sampletrees)){
  treedata_min[[tr]] <- geiger::treedata(sampletrees[[tr]],sp_data_min,sort=TRUE,warnings=F)
  treedata_max[[tr]] <- geiger::treedata(sampletrees[[tr]],sp_data_max,sort=TRUE,warnings=F)
  full_est <- nodeEstimateFossils(treedata_min[[tr]],treedata_max[[tr]])
  node_est[[tr]] <- full_est$est
}
relist1 <- lapply(lapply(1:length(sampletrees), function(x) array(unlist(node_est), 
dim = c(2, 52, 3, 100))[,,1,x]), list)
ppgmMESS(sp_data_min, sp_data_max, est=list(relist1) ,sampletrees, fossils=biofossils,
timeslice=10, which.biovars=1, which.plot="all")
}
}
\seealso{
\code{ppgm()}
}
\author{
A. Michelle Lawing, Alexandra F. C. Howard, Maria-Aleja Hurtado-Materon
}
