% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source_functions.R
\name{ppgmMESS}
\alias{ppgmMESS}
\title{ppgmMESS}
\usage{
plotMESS()
}
\arguments{
\item{cem_min}{the cem min output from the ppgm function. cbind() if there are multiple variables.}

\item{cem_max}{the cem max output from the ppgm function. cbind() if there are multiple variables.}

\item{est}{the node_est output from the ppgm function.}

\item{tree}{the phylogeny or multiple phylogenies that show the relationship between species}

\item{timeslice}{the time in million of years ago to project MESS maps (0 to 20). can handle single timeslice or vector of times.}

\item{which.biovars}{the biovariable number(s) between 1 and 19. handles vectors.}
}
\description{
This creates a MESS map for given time slices, climate envelopes, and paleoclimate models.
}

